Refactor: Extract to OffenceLocationMethods
